delete and change tests
